Reorganize code to separate library function from caller
